move some network functions around
