+++ network/tcpserver.go
@@ -13,21 +13,6 @@ const (
 	maxMsgLen = 1 << 24
 )
 
-// An Address contains the information needed to contact a peer over TCP.
-type Address string
-
-// Host returns the Address' IP.
-func (a Address) Host() string {
-	host, _, _ := net.SplitHostPort(string(a))
-	return host
-}
-
-// Port returns the Address' port number.
-func (a Address) Port() string {
-	_, port, _ := net.SplitHostPort(string(a))
-	return port
-}
-
 // A TCPServer sends and receives messages. It also maintains an address book
 // of peers to broadcast to and make requests of.
 type TCPServer struct {
